move: offer - schema
