+++ app/entities/offer/schema_offer.py
