+++ app/tests/test_offer.py
@@ -11,7 +11,7 @@ from entities.offer.models_offer import OfferLinkModel
 from tests.utils import url_reverse
 from tests.conftest import engine
 from entities.offer.models_offer import OfferLinkInSchema
-from schemas.offer_schema import OfferLinkOptionalSchema
+from entities.offer.schema_offer import OfferLinkOptionalSchema
 from routers.offer_routers import OfferFilter
